// src/app/models/AddonMDL.ts
export interface AddOnMDL {
    PackageId: number;
    AddOnPremium: number;
    PackageDetailId: number;
    AddOnTypeId: number;
    AddOnType: string;
    AddOnTenure: number;
    Action: number;
    AddOnName: string;
    PackageName: string;
    isChecked: boolean;
    IsCoverValue: boolean;
    IsDisabled: boolean;
    CoverValue: string;
    AddOnDescription: string
}